refactor(osd-viewer-demo): change example worker files name
