+++ apps/viewer-demo/src/workers/WebWorker.js
